Revert "RDKAAMP-2576 build simulator with (mocked) opencdm"
This reverts commit e0dba70de4ce76dd5bb0eddfc32ca12baedeb17e.


Revert "Revert "Revert "RDKAAMP-2576 build simulator with (mocked) opencdm"""

This reverts commit 122c4d1350be5212e36f0904fc3e3c3c3b099868.

+++ drm/AampHlsOcdmBridge.cpp
@@ -75,7 +75,7 @@ DrmReturn AampHlsOcdmBridge::Decrypt( ProfilerBucketType bucketType, void *encry
 	if (m_drmState == eDRM_KEY_ACQUIRED)
 	{
 		 AAMPLOG_TRACE("Starting decrypt\n");
-		 int retVal = m_drmSession->decrypt(m_drmInfo->iv, DRM_IV_LEN, (const uint8_t *)encryptedDataPtr , (uint32_t)encryptedDataLen, NULL);
+		 int retVal = m_drmSession->decrypt(m_drmInfo->iv, DRM_IV_LEN, (const uint8_t *)encryptedDataPtr , encryptedDataLen, NULL);
 		 if (retVal)
 		 {
 			AAMPLOG_WARN("Decrypt failed err = %d\n", retVal);
